Refactor: Move Python bindings into shared module
