+++ pyavd_utils/THIRD_PARTY_LICENSES.txt
@@ -10,8 +10,7 @@ Apache License 2.0
 Used by:
   - avdschema 0.0.7
   - passwords 0.0.7
-  - pypasswords 0.0.7
-  - pyvalidation 0.0.7
+  - python_bindings 0.0.7
   - validation 0.0.7
   - yaml-parser 0.0.7
   - ryu 1.0.23
@@ -199,7 +198,7 @@ MIT License
 
 Used by:
   - log 0.4.33
-  - regex-automata 0.4.14
+  - regex-automata 0.4.16
   - regex-syntax 0.8.11
 
 License text:
@@ -522,7 +521,7 @@ MIT License
 ================================================================================
 
 Used by:
-  - arc-swap 1.9.1
+  - arc-swap 1.9.2
   - pyo3-log 0.13.4
 
 License text:
@@ -879,7 +878,7 @@ MIT License
 ================================================================================
 
 Used by:
-  - hybrid-array 0.4.12
+  - hybrid-array 0.4.13
 
 License text:
 --------------------------------------------------------------------------------
@@ -1146,7 +1145,7 @@ MIT License
 ================================================================================
 
 Used by:
-  - simd-adler32 0.3.9
+  - simd-adler32 0.3.10
 
 License text:
 --------------------------------------------------------------------------------
@@ -1243,18 +1242,19 @@ Used by:
   - adler2 2.0.1
   - itoa 1.0.18
   - once_cell 1.21.4
-  - proc-macro2 1.0.106
-  - quote 1.0.45
-  - rustversion 1.0.22
-  - serde 1.0.228
-  - serde_core 1.0.228
-  - serde_derive 1.0.228
-  - serde_json 1.0.150
+  - proc-macro2 1.0.107
+  - quote 1.0.47
+  - rustversion 1.0.23
+  - serde 1.0.229
+  - serde_core 1.0.229
+  - serde_derive 1.0.229
+  - serde_json 1.0.151
   - serde_yaml 0.9.34+deprecated
-  - syn 2.0.117
+  - syn 2.0.119
+  - syn 3.0.2
   - unicode-ident 1.0.24
   - unsafe-libyaml 0.2.11
-  - zmij 1.0.21
+  - zmij 1.0.23
 
 License text:
 --------------------------------------------------------------------------------
@@ -1381,7 +1381,7 @@ MIT License
 
 Used by:
   - aho-corasick 1.1.4
-  - memchr 2.8.2
+  - memchr 2.8.3
   - walkdir 2.5.0
 
 License text:
