+++ pyavd_utils/passwords.py
@@ -1,0 +1,15 @@
+# Copyright (c) 2025-2026 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
+from __future__ import annotations
+
+from ._bindings import cbc_decrypt, cbc_encrypt, cbc_verify, sha512_crypt, simple_7_decrypt, simple_7_encrypt
+
+__all__ = [
+    "cbc_decrypt",
+    "cbc_encrypt",
+    "cbc_verify",
+    "sha512_crypt",
+    "simple_7_decrypt",
+    "simple_7_encrypt",
+]
