+++ pyavd_utils/schema_store.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2025-2026 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
+from __future__ import annotations
+
+from ._bindings import init_store_from_file
+
+__all__ = ["init_store_from_file"]
