--- rust/pypasswords/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "pypasswords"
-version.workspace = true
-edition = "2024"
-license.workspace = true
-
-[lints]
-workspace = true
-
-[build-dependencies]
-pyo3-build-config = { workspace = true }
-
-[dependencies]
-pyo3 = { workspace = true, "features" = ["abi3-py310"] }
-passwords = { path = "../passwords", default-features = false }
-
-[features]
-default = ["cbc", "sha512", "simple-7"]
-cbc = ["passwords/cbc"]
-sha512 = ["passwords/sha512"]
-simple-7 = ["passwords/simple-7"]
-
-[lib]
-crate-type = ["cdylib", "lib"]
