+++ rust/python_bindings/src/tests/passwords/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) 2025-2026 Arista Networks, Inc.
+// Use of this source code is governed by the Apache License 2.0
+// that can be found in the LICENSE file.
+
+#[cfg(feature = "cbc")]
+mod test_cbc;
+
+#[cfg(feature = "sha512")]
+mod test_sha512;
+
+#[cfg(feature = "simple-7")]
+mod test_simple_7;
