+++ rust/python_bindings/src/tests/passwords/test_cbc.rs
@@ -2,18 +2,19 @@
 // Use of this source code is governed by the Apache License 2.0
 // that can be found in the LICENSE file.
 
-use super::*;
+use pyo3::types::PyAnyMethods as _;
+
+use super::super::setup_python;
 
 #[test]
 fn cbc_decrypt_invalid_base64_err() {
-    setup();
+    setup_python();
     pyo3::Python::attach(|py| {
-        let module = py.import("passwords").unwrap();
-        let args = ("passwd", "ThisIsNotBase64!!!");
+        let module = py.import("_bindings").unwrap();
+        let err = module
+            .call_method1("cbc_decrypt", ("passwd", "ThisIsNotBase64!!!"))
+            .unwrap_err();
 
-        let err = module.call_method1("cbc_decrypt", args).unwrap_err();
-
-        // Maps CbcError::InvalidBase64 -> PyValueError
         assert!(err.is_instance_of::<pyo3::exceptions::PyValueError>(py));
         assert_eq!(err.value(py).to_string(), "Invalid Base64 encoding");
     });
@@ -21,13 +22,12 @@ fn cbc_decrypt_invalid_base64_err() {
 
 #[test]
 fn cbc_decrypt_failed_err() {
-    with_passwords_module(|py, module| {
-        // "YWJjZA==" is "abcd" (4 bytes).
-        // Triple DES requires multiples of 8.
-        // This will trigger the .map_err(|_| CbcError::DecryptionFailed) branch.
-        let args = ("any_key", "YWJjZA==");
-
-        let err = module.call_method1("cbc_decrypt", args).unwrap_err();
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
+        let err = module
+            .call_method1("cbc_decrypt", ("any_key", "YWJjZA=="))
+            .unwrap_err();
 
         assert!(err.is_instance_of::<pyo3::exceptions::PyRuntimeError>(py));
         assert_eq!(
@@ -39,10 +39,12 @@ fn cbc_decrypt_failed_err() {
 
 #[test]
 fn cbc_decrypt_invalid_signature_err() {
-    with_passwords_module(|py, module| {
-        // Provide valid base64, but no Arista signature at the beginning
-        let args = ("some_key", "YWFhYWFhYWFhYWFhYWFhYQ==");
-        let err = module.call_method1("cbc_decrypt", args).unwrap_err();
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
+        let err = module
+            .call_method1("cbc_decrypt", ("some_key", "YWFhYWFhYWFhYWFhYWFhYQ=="))
+            .unwrap_err();
 
         assert!(err.is_instance_of::<pyo3::exceptions::PyRuntimeError>(py));
         assert_eq!(
@@ -54,7 +56,9 @@ fn cbc_decrypt_invalid_signature_err() {
 
 #[test]
 fn cbc_verify_returns_bool() {
-    with_passwords_module(|_py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let key = "42.42.42.42";
         let data = "arista";
 
@@ -63,21 +67,18 @@ fn cbc_verify_returns_bool() {
             .unwrap()
             .extract()
             .unwrap();
-
-        // Success case
         let is_valid: bool = module
             .call_method1("cbc_verify", (key, encrypted.clone()))
             .unwrap()
             .extract()
             .unwrap();
-        assert!(is_valid);
-
-        // Failure case
         let is_invalid: bool = module
             .call_method1("cbc_verify", ("wrong_key", encrypted))
             .unwrap()
             .extract()
             .unwrap();
+
+        assert!(is_valid);
         assert!(!is_invalid);
     });
 }
