+++ rust/python_bindings/src/tests/passwords/test_simple_7.rs
@@ -2,20 +2,39 @@
 // Use of this source code is governed by the Apache License 2.0
 // that can be found in the LICENSE file.
 
-use super::*;
+use pyo3::types::PyAnyMethods as _;
+
+use super::super::setup_python;
+
+#[test]
+fn simple_7_encrypt_py_ok() {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
+        let encrypted = {
+            let args = ();
+            let kwargs = pyo3::types::PyDict::new(py);
+            kwargs.set_item("data", "arista").unwrap();
+            kwargs.set_item("salt", 4).unwrap();
+            module
+                .call_method("simple_7_encrypt", args, Some(&kwargs))
+                .unwrap()
+        };
+        assert_eq!(encrypted.to_string(), "045A190F1C354D");
+    });
+}
 
 #[test]
 fn simple_7_encrypt_decrypt_roundtrip() {
-    with_passwords_module(|_py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let password = "test_password";
-        let salt = 5_u8;
-
         let encrypted: String = module
-            .call_method1("simple_7_encrypt", (password, Some(salt)))
+            .call_method1("simple_7_encrypt", (password, Some(5_u8)))
             .unwrap()
             .extract()
             .unwrap();
-
         let decrypted: String = module
             .call_method1("simple_7_decrypt", (encrypted,))
             .unwrap()
@@ -28,17 +47,15 @@ fn simple_7_encrypt_decrypt_roundtrip() {
 
 #[test]
 fn simple_7_encrypt_with_random_salt() {
-    with_passwords_module(|py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let password = "test_password";
-
-        // Call with None for salt
         let encrypted: String = module
             .call_method1("simple_7_encrypt", (password, py.None()))
             .unwrap()
             .extract()
             .unwrap();
-
-        // Should be able to decrypt it
         let decrypted: String = module
             .call_method1("simple_7_decrypt", (encrypted,))
             .unwrap()
@@ -51,7 +68,9 @@ fn simple_7_encrypt_with_random_salt() {
 
 #[test]
 fn simple_7_encrypt_empty_password_err() {
-    with_passwords_module(|py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let err = module
             .call_method1("simple_7_encrypt", ("", Some(5_u8)))
             .unwrap_err();
@@ -63,15 +82,13 @@ fn simple_7_encrypt_empty_password_err() {
 
 #[test]
 fn simple_7_encrypt_invalid_salt_err() {
-    with_passwords_module(|py, module| {
-        let password = "test_password";
-        let invalid_salt = 16_u8; // Out of range (0-15)
-
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let err = module
-            .call_method1("simple_7_encrypt", (password, Some(invalid_salt)))
+            .call_method1("simple_7_encrypt", ("test_password", Some(16_u8)))
             .unwrap_err();
 
-        // Maps Simple7Error::InvalidSaltValue -> PyValueError
         assert!(err.is_instance_of::<pyo3::exceptions::PyValueError>(py));
         assert_eq!(
             err.value(py).to_string(),
@@ -82,7 +99,9 @@ fn simple_7_encrypt_invalid_salt_err() {
 
 #[test]
 fn simple_7_decrypt_data_too_short_err() {
-    with_passwords_module(|py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let err = module.call_method1("simple_7_decrypt", ("0",)).unwrap_err();
 
         assert!(err.is_instance_of::<pyo3::exceptions::PyValueError>(py));
@@ -95,7 +114,9 @@ fn simple_7_decrypt_data_too_short_err() {
 
 #[test]
 fn simple_7_decrypt_invalid_hex_err() {
-    with_passwords_module(|py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let err = module
             .call_method1("simple_7_decrypt", ("01GGGG",))
             .unwrap_err();
@@ -107,7 +128,9 @@ fn simple_7_decrypt_invalid_hex_err() {
 
 #[test]
 fn simple_7_decrypt_invalid_salt_format_err() {
-    with_passwords_module(|py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let err = module
             .call_method1("simple_7_decrypt", ("XX1234",))
             .unwrap_err();
@@ -119,7 +142,9 @@ fn simple_7_decrypt_invalid_salt_format_err() {
 
 #[test]
 fn simple_7_decrypt_salt_out_of_range_err() {
-    with_passwords_module(|py, module| {
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let err = module
             .call_method1("simple_7_decrypt", ("161234",))
             .unwrap_err();
@@ -134,8 +159,9 @@ fn simple_7_decrypt_salt_out_of_range_err() {
 
 #[test]
 fn simple_7_known_values() {
-    with_passwords_module(|_py, module| {
-        // Test known encryption values
+    setup_python();
+    pyo3::Python::attach(|py| {
+        let module = py.import("_bindings").unwrap();
         let test_cases: [(u8, &str, &str); 4] = [
             (1, "foo", "0115090B"),
             (6, "foo", "0600002E"),
@@ -149,13 +175,13 @@ fn simple_7_known_values() {
                 .unwrap()
                 .extract()
                 .unwrap();
-            assert_eq!(encrypted, expected_encrypted);
-
             let decrypted: String = module
                 .call_method1("simple_7_decrypt", (expected_encrypted,))
                 .unwrap()
                 .extract()
                 .unwrap();
+
+            assert_eq!(encrypted, expected_encrypted);
             assert_eq!(decrypted, password);
         }
     });
