+++ rust/test_schema_store/Cargo.toml
@@ -2,6 +2,8 @@
 name = "test_schema_store"
 edition = "2024"
 
+[workspace]
+
 [dependencies]
 reqwest = { version = "0.12.24", default-features = false, features = [
     "blocking",
